CA-157 Not able to introduce a node/location lying on existing segment
Had to implement the ON_SEGMENT recommendation along with confirming that
recommendation.  Got into a bit of outstanding refactoring work too.
