+++ crCore/src/main/java/com/clueride/service/DefaultNodeService.java
@@ -46,11 +46,6 @@ import com.clueride.domain.dev.NodeGroup;
 import com.clueride.domain.dev.NodeNetworkState;
 import com.clueride.domain.dev.rec.DiagnosticRec;
 import com.clueride.domain.dev.rec.Rec;
-import com.clueride.domain.dev.rec.ToNode;
-import com.clueride.domain.dev.rec.ToSegment;
-import com.clueride.domain.dev.rec.ToSegmentAndNode;
-import com.clueride.domain.dev.rec.ToTwoNodes;
-import com.clueride.domain.dev.rec.ToTwoSegments;
 import com.clueride.domain.factory.PointFactory;
 import com.clueride.feature.Edge;
 import com.clueride.feature.LineFeature;
@@ -97,209 +92,6 @@ public class DefaultNodeService implements NodeService {
         return proposalSummary.toJson();
     }
 
-    /**
-     * Implementation of a request to confirm the latest proposal as the one we
-     * want to add to the network.
-     *
-     * The NetworkProposal instance we stashed away will hold the details needed
-     * to create the objects we add to the persisted network.
-     *
-     * @param recId - Integer representing the unique recommendation to be committed
-     *              to the network.
-     * @return upon success, returns JSON status:OK.
-     */
-    @Override
-    public String confirmRecommendation(Integer recId) {
-        NetworkProposal networkProposal = NetworkProposalStore
-                .getLastProposal();
-        Rec rec = (Rec) networkProposal.getRecommendation(recId);
-        if (rec != null) {
-            switch (rec.getRecType()) {
-                case ON_SEGMENT:
-                case ON_NODE:
-                    // TODO: May become part of the signature for a RecommendationService
-                    LOGGER.warn("We don't have a way to select "+rec.getRecType());
-                    break;
-
-                case TRACK_TO_NODE:
-                    addTrackToNodeRec((ToNode) rec);
-                    break;
-                case TRACK_TO_SEGMENT:
-                    addTrackToSegmentRec((ToSegment) rec);
-                    break;
-                case TRACK_TO_SEGMENT_AND_NODE:
-                    addTrackToSegmentAndNodeRec((ToSegmentAndNode) rec);
-                    break;
-                case TRACK_TO_2_SEGMENTS:
-                    addTrackToTwoSegmentsRec((ToTwoSegments) rec);
-                    break;
-                case TRACK_TO_2_NODES:
-                    addTrackToTwoNodesRec((ToTwoNodes) rec);
-                    break;
-
-                case UNDEFINED:
-                case OFF_NETWORK:
-                default:
-                    LOGGER.warn("Rec Type: " + rec.getRecType());
-            }
-        }
-        DefaultNetwork.getInstance().storesReadyForPublishing();
-        return "{\"status\": \"OK\"}";
-    }
-
-    /**
-     * Prepares the following items for adding to the network:
-     * <ul>
-     *     <li>The New Node itself.</li>
-     *     <li>The Edge connecting the Node to the Network Node.</li>
-     * </ul>
-     *
-     * These are then persisted.
-     * @param rec - ToNode instance containing the New Node and the new Edge
-     *            which connects the New Node to the Network Node.
-     */
-    private void addTrackToNodeRec(ToNode rec) {
-        LOGGER.info("From this Rec: "+rec);
-        LOGGER.info("Preparing the following pieces to add to the Network:");
-        LOGGER.info("New Loc: " + rec.getNewNode().getName());
-        rec.logRecommendationSummary();
-
-        Integer nodeId = nodeStore.addNew(rec.getNewNode());
-        LOGGER.info("New Node with ID: " + nodeId);
-
-        SegmentService.addSegment(rec.getProposedTrack());
-    }
-
-    /**
-     * Prepares these items for adding to the network:
-     * <ul>
-     * <li>The Location Node itself
-     * <li>The Edge connecting the Location to the Segment (from the Track)
-     * <li>The existing Network Edge/Segment where the proposed Edge meets the network.</li>
-     * <li>The Splitting Node (also the end of the previous Edge).
-     * </ul>
-     * We also tell the SegmentService about the existing Network Edge and the
-     * location where we join in so the Edge can be split into two Edges creating
-     * a three-way intersection at that Splitting Node.
-     * These are then persisted.
-     * @param rec - Containing specific elements to be added to the Network.
-     */
-    private void addTrackToSegmentRec(ToSegment rec) {
-        LOGGER.info("From this Rec: " + rec);
-        LOGGER.info("Preparing the following pieces to add to the Network:");
-
-        // Two Nodes:
-        Integer nodeId = nodeStore.addNew(rec.getNewNode());
-        LOGGER.info("New Node with ID: " + nodeId);
-        Integer splittingNodeId = nodeStore.addNew(rec.getSplittingNode());
-        LOGGER.info("Splitting Node with ID: " + splittingNodeId);
-
-        // Accepted Proposal for Edge/Segment from new node to the existing segment
-        SegmentService.addSegment(rec.getProposedTrack());
-
-        // Split of the Existing Network Node
-        SegmentService.splitSegment(rec.getSegment(), rec.getSplittingNode());
-        rec.logRecommendationSummary();
-    }
-
-    /**
-     * Prepares the following items for adding to the network:
-     * <ul>
-     *     <li>The New Node itself.</li>
-     *     <li>The Edge in one direction connecting the Node to a Network Node or Edge.</li>
-     *     <li>The Edge in the other direction connecting the new Node to either a Network Node or Edge.</li>
-     *     <li>The existing Network Edge/Segment where the proposed Edge meets the network.</li>
-     *     <li>The Splitting Node where the proposed Edge meets the Network Edge.
-     * </ul>
-     * We also tell the SegmentService about the existing Network Edge and the
-     * location where we join in so the Edge can be split into two Edges creating
-     * a three-way intersection at that Splitting Node.
-     * @param rec - Containing specific elements to be added to the Network.
-     */
-    private void addTrackToSegmentAndNodeRec(ToSegmentAndNode rec) {
-        LOGGER.info("From this Rec: "+rec);
-        LOGGER.info("Preparing the following pieces to add to the Network:");
-
-        // Two Nodes:
-        Integer nodeId = nodeStore.addNew(rec.getNewNode());
-        LOGGER.info("New Node with ID: " + nodeId);
-        Integer splittingNodeId = nodeStore.addNew(rec.getSplittingNode());
-        LOGGER.info("Splitting Node with ID: " + splittingNodeId);
-
-        // Accepted Proposal for Edge/Segment from new node to the existing segment
-        SegmentService.addSegment(rec.getProposedTracks().get(0));
-        SegmentService.addSegment(rec.getProposedTracks().get(1));
-
-        // Split of the Existing Network Node
-        SegmentService.splitSegment(rec.getSegment(), rec.getSplittingNode());
-        rec.logRecommendationSummary();
-    }
-
-    /**
-     * Prepares these items for adding to the network:
-     * <ul>
-     *     <li>The New Node itself.</li>
-     *     <li>A new Edge in one direction from the new Node over to a Network Edge.</li>
-     *     <li>A new Edge in the other direction from the new Node over to a Network Edge.</li>
-     *     <li>The Splitting Node where the first Edge hits the Network</li>
-     *     <li>The Splitting Node where the second Edge hits the Network</li>
-     * </ul>
-     * These are then persisted.
-     * We also tell the SegmentService about the two existing Network Edges and the
-     * location where we join in so those Edges can be split into two Edges creating
-     * a three-way intersection at that Splitting Node.
-     * @param rec - ToTwoSegments recommendation containing New Node, the two Edges
-     *            leading in either direction toward two other existing Network Edges
-     *            and the two Splitting Nodes, one for each Network Edge to be split.
-     */
-    private void addTrackToTwoSegmentsRec(ToTwoSegments rec) {
-        LOGGER.info("From this Rec: "+rec);
-        LOGGER.info("Preparing the following pieces to add to the Network:");
-
-        // Two Nodes:
-        Integer nodeId = nodeStore.addNew(rec.getNewNode());
-        LOGGER.info("New Node with ID: " + nodeId);
-        Integer splittingNodeEndId = nodeStore.addNew(rec.getSplittingNodeEnd());
-        LOGGER.info("Splitting Node with ID: " + splittingNodeEndId);
-        Integer splittingNodeStartId = nodeStore.addNew(rec.getSplittingNodeStart());
-        LOGGER.info("Splitting Node with ID: " + splittingNodeStartId);
-
-        // Accepted Proposal for Edge/Segment from new node to the existing segment
-        SegmentService.addSegment(rec.getProposedTracks().get(0));
-        SegmentService.addSegment(rec.getProposedTracks().get(1));
-
-        // Split of the Existing Network Node
-        SegmentService.splitSegment(rec.getEndSegment(), rec.getSplittingNodeEnd());
-        SegmentService.splitSegment(rec.getStartSegment(), rec.getSplittingNodeStart());
-
-        rec.logRecommendationSummary();
-    }
-
-    /**
-     * Prepares the following items for adding to the network:
-     * <ul>
-     *     <li>The New Node itself.</li>
-     *     <li>The Edge in one direction connecting the Node to the Network Node.</li>
-     *     <li>The Edge in the other direction connecting the Node to another Network Node.</li>
-     * </ul>
-     * These are then persisted.
-     * @param rec - Containing specific elements to be added to the Network.
-     */
-    private void addTrackToTwoNodesRec(ToTwoNodes rec) {
-        LOGGER.info("From this Rec: " + rec);
-        LOGGER.info("Preparing the following pieces to add to the Network:");
-
-        /* New Node: */
-        Integer nodeId = nodeStore.addNew(rec.getNewNode());
-        LOGGER.info("New Node with ID: " + nodeId);
-
-        /* Accepted Proposal for Edge/Segment from new node to the existing node in each direction. */
-        SegmentService.addSegment(rec.getProposedTracks().get(0));
-        SegmentService.addSegment(rec.getProposedTracks().get(1));
-
-        rec.logRecommendationSummary();
-    }
-
     /* Node Group Endpoint. */
 
     @Override
@@ -327,12 +119,6 @@ public class DefaultNodeService implements NodeService {
         LOGGER.info("Requesting All Nodes");
         NetworkProposal networkProposal = buildAllNodesProposal();
         return networkProposal.toJson();
-    }
-
-    @Override
-    public String getRecGeometry(Integer recId) {
-        LOGGER.info("Requesting Geometry for Rec ID: " + recId);
-        return recommendationService.getRecGeometry(recId);
     }
 
     @Override
