+++ crCore/src/main/java/com/clueride/service/NodeService.java
@@ -49,19 +49,6 @@ public interface NodeService {
 
     String setNodeGroup(Integer id, Double lat, Double lon);
 
-    // TODO: Probably goes over to the RecommendationService
-    String getRecGeometry(Integer recId);
-
-    /**
-     * Accept specific recommendation from the Proposal identified by the ID.
-     * @param recId - unique ID of the recommendation to confirm.
-     * @return the string "OK" to indicate success.
-     */
-    // TODO: Probably goes over to the RecommendationService
-    String confirmRecommendation(Integer recId);
-
-
-
     // TODO: Probably would go over to the EdgeService
     String getMatchingSegments(Integer pointId);
