+++ crCore/src/main/java/com/clueride/service/builder/RecommendationBuilder.java
@@ -17,12 +17,18 @@
  */
 package com.clueride.service.builder;
 
+import com.vividsolutions.jts.geom.LineString;
+
 import com.clueride.domain.GeoNode;
 import com.clueride.domain.dev.NetworkRecommendation;
-import com.clueride.domain.dev.rec.*;
+import com.clueride.domain.dev.rec.OffNetworkImpl;
+import com.clueride.domain.dev.rec.OnNodeImpl;
+import com.clueride.domain.dev.rec.OnSegmentImpl;
+import com.clueride.domain.dev.rec.ToNodeImpl;
+import com.clueride.domain.dev.rec.ToSegmentImpl;
+import com.clueride.domain.dev.rec.ToTwoNodesImpl;
 import com.clueride.feature.Edge;
 import com.clueride.feature.TrackFeature;
-import com.vividsolutions.jts.geom.LineString;
 
 /**
  * Builder for a Recommendation that becomes part of a Proposal.
@@ -33,6 +39,7 @@ import com.vividsolutions.jts.geom.LineString;
  * At this time, we're implementing {@link NetworkRecommendation}.
  * 
  * @author jett
+ * @deprecated in favor of TrackRecBuilder and OnSegmentRecBuilder.
  */
 public class RecommendationBuilder {
     /** The requested node which we're making a recommendation for. */
